+++ src/Client/headers/SensorController.h
@@ -1,7 +1,7 @@
 #ifndef SENSORCONTROLLER_H   /* Include guard */
 #define SENSORCONTROLLER_H
 
-int getDistanceSensorValue();
+float getDistanceSensorValue();
 
 
 #endif
